+++ src/package/themes/default/index.ts
@@ -17,6 +17,7 @@ import autocomplete from './components/autocomplete';
 import multiselect from './components/multiselect';
 import datepicker from './components/datepicker';
 import daterangepicker from './components/daterangepicker';
+import timepicker from './components/timepicker';
 import pagination from './components/pagination';
 import checkbox from './components/checkbox';
 import radio from './components/radio';
@@ -57,6 +58,7 @@ export default {
         multiselect,
         datepicker,
         daterangepicker,
+        timepicker,
         pagination,
         checkbox,
         radio,